requirements.txt
asgiref==3.7.2

beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-modeltranslation==0.18.11
export==0.2.1
feedparser==6.0.10
idna==3.4
lxml==4.9.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rss-parser==1.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
xmltodict==0.13.0
